requirements.txt
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
fastapi==0.110.0
graphql-core==3.2.3
greenlet==3.0.3
h11==0.14.0
idna==3.6
iniconfig==2.0.0
packaging==24.0
pluggy==1.4.0
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
strawberry-graphql==0.220.0
typing_extensions==4.10.0
uvicorn==0.28.0
